cosmos/detection/src/requirements.txt
numpy==1.15.4
pyyaml
hyperyaml
scikit-image
bitmath
tensorboardX
tensorboard
tqdm
opencv-python
pascal-voc-writer
imgaug
dominate
pytesseract
Pillow
redis
click
sqlalchemy
pandas
dataclasses
